docker_test/requirements.txt
cachetools==2.0.1
certifi==2018.4.16
chardet==3.0.4
click==6.7
Flask==0.12.2
futures==3.2.0
google-api-core==1.1.1
google-auth==1.4.1
google-cloud-core==0.28.1
google-cloud-storage==1.8.0
google-resumable-media==0.3.1
googleapis-common-protos==1.5.3
idna==2.6
itsdangerous==0.24
Jinja2==2.10
MarkupSafe==1.0
protobuf==3.5.2.post1
pyasn1==0.4.2
pyasn1-modules==0.2.1
pytz==2018.4
requests==2.18.4
rsa==3.4.2
six==1.11.0
urllib3==1.22
Werkzeug==0.14.1
